chore: 스토리북 설정
